results/rebuttal_new_baselines/run-20231112-040536_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_20-runs_log/dropbox/4/CodeT/test_file_management.py
import pytest
import file_management

def test_upload():
	data = {'name': 'Test', 'size': 10, 'type': 'txt', 'versions': []}
	response = file_management.upload(data)
	assert response == {'message': 'File uploaded successfully'}


def test_download():
	data = {'name': 'Test'}
	response = file_management.download(data)
	assert 'name' in response


def test_organize():
	data = {'name': 'Test', 'new_name': 'NewTest'}
	response = file_management.organize(data)
	assert response == {'message': 'File organized successfully'}


def test_versioning():
	data = {'name': 'Test'}
	response = file_management.versioning(data)
	assert 'versions' in response
